Fix cases where console_url and server_url were missing (#2528)
When using Pydantic, a Field alias will be used to override the name of
a field when it's initialized. In order to use both the field name and
alias, which is useful when dealing with snake case (Python) and camel
case (JSON), the allow_population_by_field_name config can be used.