[jenkins-job-builder] reduce required schemas for integration run
the jenkins-job-builder integration queries a lot more schemas than required because it uses some generic app-interface-settings and aws-account queries th
at are to broad.

this change introduces a `get_secret_reader_settings()` query that fetches minimal information for secret reader initialization.
also the `State` can be initialized with a new function `init_state` that requires minimal data about the involved account.

as a result, a lot of schemas can be removed from the jenkins-jobs-builders `integration-1.schemas` list in app-interface, which has positive side effects for PR checks. also the `integration-1.pr_check.run_for_valid_saas_file_changes` option can be removed for this integration, since the saas-file-2 schema is not required to be in the schema list anymore. the remaining schemas are

```yaml
schemas:
- /app-sre/integration-1.yml
- /app-interface/app-interface-settings-1.yml # for secret reader
- /dependencies/jenkins-config-1.yml # for jobs
- /dependencies/jenkins-instance-1.yml # for jobs
- /aws/account-1.yml # for state
- /app-sre/app-1.yml # for code components
- /access/user-1.yml # for crosscheck repo admins
- /access/external-user-1.yml # for crosscheck repo admins
- /access/bot-1.yml # for crosscheck repo admins
```

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>