Locally scope cli.py imports to save memory usage
Every import in cli.py has to complete before any integration begins
running. This is because import statements are executed when the cli
module is imported, which happens as a result of console_scripts in
setup.py. Currently, memory usage can reach ~120MB before an integration
ever starts running. This change scopes the imports inside the integration
command functions so that the import only happens when the function
is executed (we only expect to execute a single integration function per
container).

There are probably more elegant ways to fix this issue, but for now
this should work and saves us 50MB or more of imports for many
integrations that don't require the importing of the majority of
qontract-reconcile modules.