Refactor unleash implementation (#2596)
* Refactor unleash implementation

Keep the client, do not create one per Request. This reduces the number
of requests to unleash. This reduces likelyhood of DNS errors due to
failed lookups during cluster upgrades.

Related ticket: APPSRE-5719

* Add test

Add a test, that outlines the multithreading capabilities of this unleash implementation.

* Change _get_unleash_api_client

Create one instance of UnleashClient, that is shared