Unleash client refactor (#2616)
* Refactor unleash implementation

Keep the client, do not create one per Request. This reduces the number
of requests to unleash. This reduces likelyhood of DNS errors due to
failed lookups during cluster upgrades.

Related ticket: APPSRE-5719

* Add test

Add a test, that outlines the multithreading capabilities of this unleash implementation.

* Change _get_unleash_api_client

Create one instance of UnleashClient, that is shared

* Fix logging

Remove logger toggle and replace it with logger configuration, that sets the log level for the UnleashClient and the JobScheduler to error. By this log messages from the UnleashClient will be skipped

* Move setLoglevel

Move setting loglevel in front of UnleashClient init. Makes sure to supress evey logmessage