Remove managed resource names from tkn integration (#2621)
This was wise while migrating to avoid clashes between resources from
openshift-resources and tkn integration. It is no longer necessary and
it just leaves orphaned Pipelines and Tasks from deleted or moved saas
files. If we want to manage Pipelines or Tasks from a namespace we can
still do it as the caller pattern will assure that objects created from
different integrations are only managed by the caller integration unless
there's a name clash.

part of APPSRE-6031.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>