early exit for integrations when no change in desired state
if a cli flag `--early-exit-compare-sha` provides a bundle sha, integrations supporting
early exit will fetch the desired state from the comparison and the current bundle and
compare them. if no difference an be found, the integration exits early.

this early exit feature is supposed to improve speed for PR checks. early exit will only
be considered if the integration has been started in dry-run mode.

an integration supports early exit, if the integration module provides a function

```
def early_exit_desired_state(*args, **kwargs) -> dict[str, Any]
```

it is crucial that the `early_exit_desired_state` leverages the same functionality/code
to fetch the desired state, as the `run` method does, in order to prevent a drift
between those two.

if any error occures during the desired state fetching or the comparison, the integration
will not exit early and will instead proceed with its regular execution.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>