Temp fix to handle the comparation of OCI images
sretoolbox currently doesn't handle well the comparation of OCI images
with content-type 'application/vnd.oci.image.index.v1+json'. These are
manifest list type of images that don't have `mediaType` in their
manifest so sretoolbox logic fails.

The real fix is in sretoolbox.container.image module, but that's going
to take a bit longer since we will need to change how image comparation
is done as it will need to inspect the manifest headers instead of the
manifest itself.

In the meantime, we have this to avoid the integration to crash.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>