Add integration spec override
Make it possible to set image tags  for a shard. Thus making it possible
to gradually roll out an updated image. Shards can be specified by
either supplying an AWS account or a shard id.

Will add the imageRef as an attribute to the $shard value in the helm
template. If set, that imageRef is prefered over the IMAGE_TAG
environment variable.