detect the affected shard for an openshift resource change in an MR
if an MR only changes openshift resources in a single cluster (shard), the integrations `openshift-resources`, `openshift-routes`, `openshift-vault-secrets` and `prometheus-rule-tester` can be run on that shard alone, which is A LOT faster than processing all clusters.

the general idea is, that an integration can provide a function `run_with_desired_state_diff` which is called instead of the integrations `run` function if a diff in desired state has been detected. the integration can then inspect the provided diff and decide if it is worth running shareded mode or not.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>