gitlab-housekeeper fix check for unmergable MRs
the previous attempt to check for unmergable MRs failed because the attribute `merge_error` is not present all the time in MR objects. additionally the place to check is the `get_merge_request` function.

this change removes the check for `merge_error` in favour of additional `merge_status` check

fixes https://github.com/app-sre/qontract-reconcile/pull/2745

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>