Fix compare-tags (aka deep-sync) mode (#2757) (#2763)
Compare tags (previously known as deep-sync) control file was updated
at the beginning of the run. Hence, even if the run was unsuccesful,
the next one would have seen that the previous execution of the
integration was a compare-tags one and would run in non-compare-tags
mode, which would mask any potential issue with the slow mode. This
patch fixes by changing the moment of the update of the control file to
the end of the execution of the integration.

In order to be able to properly test these changes without needing to
modify the source code, a few options have been added to the
integration to control:

* where the control file is created.
* how often to compare tags.
* to force the comparation of tags.
* to specify a certain image.

The first one is particularly interesting as it can allow us to create
the control file outside the container, which can allow us to avoid a
compare tags run every time the pod is restarted/recreated.

This was first attempted in #2757 but I revert it as builds were
failing.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>