[qontract-cli] Make template usable by non app-sre
Added local secret config support and a way to filter results by path
that will avoid the need to render templates to inspect their names.
This will make calling Vault not needed for many templates, easying the
use for tenants and making it considerably faster for everyone.

The new option --path may seem redundant but it has benefits and we are
keeping backwards compatibility just in case.

Signed-off-by: Rafa Porres Molina rporresm@redhat.com