[change-owners] deal with new or deleted files
if files were added or deleted entirely, diff detection would fail in the case of resourcefiles. this change makes sure this does not happen.

as a nice sideeffect, the old and new colums of a detected diff now show more reliable output in the case of added or removed dict elements.

additionally the diff detection part of the change-owners integration has been moved into the entry point methods try/except block so that PR checks won't fail for now while the diffing code is still stabilizing.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>