Enable overrides per shard (#2725)
* Add integration spec override

Make it possible to set image tags  for a shard. Thus making it possible
to gradually roll out an updated image. Shards can be specified by
either supplying an AWS account or a shard id.

Will add the imageRef as an attribute to the $shard value in the helm
template. If set, that imageRef is prefered over the IMAGE_TAG
environment variable.

* Add tests, fix lint/typing

* Make nested loops simpler

* Add typing to initialize_namespace_override_mapping

* Improve typing

Add more specific type for the override mapping
Fix types supplied to methods

* Add integration imageref

Add integration imageRef, making it possible to set a default imageRef
for shards, that are not explicitly defined

* Fix tests

* Fix typing and format

* Make test simpler

Remove not required test data from test

* Update query

Implemented using interfaces

* Add helm specific test

Adding a test, that ensures imageRef is handeld by the helm template

* Clean up dummy file