move diff extraction logic and datastructures in dedicated module
everything related to diff detection has been moved to reconcile/changeowners/diff.py

this implied also structurally detach a diff from the ChangeTypeContext that covers it.
instead of having Diff.covered_by, a new dataclass DiffCoverage has been introduced, that
carries a reference to a diff and a list to the coering ChangeTypeContexts.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>