Add google chat as fluentd log store option (#2801)
This MR updates the helm template to include a conditional google chat log store option. This option is used in FedRAMP environment for chatops.

This MR also adds a new values file, values-manager-fedramp.yaml. Because slack is prohibited due to FR controls, we must have a way to disable slack reporting for all integrations. Every downstream integration can manage this on its own, but integrations-manager itself must be tempalted with slack: false and googleChat: true.

The http plugin with fluentd - I spent many days with this trying to get json payloads formatted properly. Did not have much luck.

The fluent-plugin-google-chat requires gcp account access - not available in FedRAMP and also this level of complexity is not needed
https://github.com/treasuryspring/fluent-plugin-google-chat

This left me to search for other fluentd plugins which could be used in a generalized webhook api fashion - after review the microsoft teams one is doing the trick, and is included in the latest app-sre/fluentd image.