[change-owners] best effort parse resource files without schema
resource files without schema are not parsed right now, which means that fine jsonpath expressions to pinpoint the change in a file will not work in those situations.

a lot of resource files are parsable though even without a schema, e.g. ConfigMaps while not all resource files with a schema are parsable when they include jinja2 templating.

this change will try to best-effort parse resource files and falls back to handling the file content as clob if not parsable.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>