Allow tekton timeouts bigger than 60 minutes
Tekton default timeout for PipelineRuns are 60 minutes. We may need to
allow for larger timeouts for certain jobs, but we don't want to allow
for smaller timeouts since the current implementation is to delete the
underlying pods, which gets us without logs.

schema: https://github.com/app-sre/qontract-schemas/pull/312

Part of APPSRE-4670

Signed-off-by: Rafael Porres Molina <rporres@gmail.com>