[change-owners] support templating and gql queries for change- and context selectors
Lets say we want to fetch the paths of all clusters involved in a saas file. We would write a `GQL` query like this

```gql
{
    saas_files_v2(path: "/service/saas-file.yml") {
        resourceTemplates {
            targets {
                namespace {
                    cluster {
                        path
                    }
                }
            }
        }
    }
}
```

and then we could use a `jsonpath` expression to extract the cluster paths:  `saas_files_v2.resourceTemplates[*].targets[*].namespace.cluster.path`.

The interesting part is, that the `jsonpath` expression and the `GQL` query in this example are very similar to each other, because we are only interested in a leaf field of the `GQL` results. if we rewrite the `jsonpath` expression to `saas_files_v2[?(@.path=='/service/saas-file.yml')].resourceTemplates[*].targets[*].namespace.cluster.path` we would have a very concise syntax to describe the required `GQL` query and the extraction mechanism for the desired field.

Say hi to the `gql+jsonpath` expression.

When an app-interface file changes but the file that is owned is another one, a `change-types` can use the `changes.context` configuration section to define how we can find the owned file in the changed one. E.g. if a role is added to a user, the users `roles` section changes but the file that is owned by app-interface tenants is the affected `/access/role-1.yml` one. The `changes.context.selector` configuration option allows us to define with a jsonpath expression that the context for a changed user is not the changed user but the role that is added to the user, e.g. `changes.context.selector: 'roles[*].'$ref'`.

This works well for situation where the owned file can be found within the changed file. Sadly that is not always the case. Consider the situation when the owned file is an `/openshift/cluster-1.yml` but the changed file is an `/app-sre/saas-file-2.yml`. The cluster can't be found within the saas file. Instead we need to follow the resource template targets, through their namespaces towards their clusters. A hard task for `jsonpath` expressions, because they can't traverse documents, but an easy one for `GraphQL queries` where we can traverse references and even synthetic back-references with ease.

Hence, pure `jsonpath` expressions are not sufficient in this case and it would be desirable to have `GraphQL` capabilities to describe how the context for a changed file can be found.

`gql+jsonpath` to the rescue. By allowing `gql+jsonpath` as `changes.context.selector`, we can bridge that gap between files and find context where it was not possible before.

Lets say we want to allow owners of a cluster to approve changes to saas file targets on namespaces that belong to their cluster, while they don't own the saas file itself, e.g. central AppSRE saas files for cert-manager deployments. A change-type should be able to grant self-service permission to deploy `cert-manager` on a cluster owner, without granting permissions on any other parts of the saas file that is not related to namespace of their cluster.

First we need to establish context, basically describing in what situations the change-type would grant access to the saas file. We can do this with the following `gql+jsonpath` example.

```yaml
contextSchema: /openshift/cluster-1.yml
changes:
- provider: jsonPath
  ...
  changeSchema: /app-sre/saas-file-2.yml
  context:
    selector: gql+jsonpath://saas_files_v2[?(@.path=='{{ changed_file_path }}')].resourceTemplates[*].targets[*].namespace.cluster.path
    when: added
  ...
```

If we can reach a cluster we own from the saas file that changed, we successfully established context.

Next we need to restrict what exactly we can change in the saas file. Obviously we should not be able to change whatever target we want but we should be restricted to the ones that reference namespaces that belong to our cluster. Instead of using static `changes.jsonPathSelector` to describe what we can change, we can use the newly introduced `changes.jsonPathSelectorTemplates` option that allows us to generate a set of `jsonpath` expressions based on a `GraphQL` query defined in `gql+jsonpath`.

```yaml
changes:
- provider: jsonPath
  ...
  changeSchema: /app-sre/saas-file-2.yml
  jsonPathSelectorTemplates:
  - var: ns
    items: gql+jsonpath://clusters_v1[?(@.path=='{{ ctx_file_path }}')].namespaces[*].path
    template: resourceTemplates[*].targets[?(@.namespace.'$ref'=='{{ ns }}')]
```

`ctx_file_path` descibes the established context, our owned cluster, hence we will fetch a list of all namespace paths from that cluster and use each one as variable `ns` to create a list of jsonpath expressions from the `template`.

This way we successfully restricted change permissions on those targets referencing namespaces from our owned cluster.

schema change - https://github.com/app-sre/qontract-schemas/pull/331

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>