[change-owners] split high level changes into fine grained diffs (#3058)
When the detected changes within a file are too rough/high level, only high level `change-types` would be able to cover them. e.g. if a new file is introduced, the detected change would be the entire file, only coverable by the top-level JSONPath placeholder for the file-root '$'. Such `change-types` are the opposite of what we intend to do with `fine-grained` permissions.

Split high level changes in smaller changes so they can be covered by `change-types`. The process of splitting is driven by the involved change-types, so that the resulting splits are as fine as they are required to allow the change-types to have effect, but not finer than required in order to keep the list of changes as concise as possible.

The center-piece of this change is the list of splits (`_split_into`) within the `DiffCoverage` dataclass which are `DiffCoverage` objects themselves. This forms a tree-like structure of diffs/sub-diffs, with the most fine grained splits represented as leaves. A diff is considered covered, if it is directly covered by a `change-type` or if all the splits of the diff are covered.

The approval process does not know anything about this process since it will operate on a flattened and consolidated list of diffs as if they would have been detected that way to begin with.

Ref: https://issues.redhat.com/browse/APPSRE-6606