refactor: refactor ownership context detection
introduce an OwnershipContext class to detect proper ownership context with `selector` and `when` conditions.

this can be reused later when we put change-types into new ownership contexts.

additionally a dataclass `ChangeDetector` has been introduced to represent the `changes` items of a `ChangeType`. currently it has only one subclass `JsonPathChangeDetector` but when we introduce the `change-type` provider for change records, we will introduce a second one.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>