refactor: introduce `ResolvedContext` class to represent a context
before this, a resolved context was represented only as a `FileRef`.
it is valueable though to track more details, e.g. how was the context resolved.

this makes sense in the following commits, when contexts result from change-type composition and in such a case we would love to track the change-type that introduced the context.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>