[integration-manager] sharding labels/matchlabels for integration workload
the current `matchLabels` on a deployment are not sufficient to catch only the pods spawned by replicasets of the deployment itself, because the `app` label alone is the same for all potential shards of an integration.

this has no negative effect because the ReplicaSets still handle pod creation/termination correctly. the only downside is the incorrect pod filtering within the openshift UI within Deployment/Pods.

to fix this, we will add a shard label `qontract_reconcile_integration_shard` to both the pod template and the matchLabels of the deployment if applicable.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>