reusable change-types through composition (#3105)
Add a new change provider named change-type to /app-interface/change-type-1.yml#changes.provider that can reference an existing change-type and put it into the context of changes.context.

```yaml
name: app-owner
...
changes:
- provider: change-type         <-- new provider
  changeTypes:
  - $ref: namespace-owner.yml   <-- puts the existing change-type ...
  context:                      <-- ... into a new context
    selector: app.'$ref'
```

This makes the referenced change-type usable in the ownership context of the referencing change-type. It redurced ownership management toil and makes the change-types reusable. It also enables the definition of higher level concepts for ownership, e.g. the owned entity is an app and by putting existing change-types into the context of the app change-type, ownership is dynamically expanded to other entities.