[change-owners] show approver groups on merge requests
this change lists approver group contact information as part of the MR change report. this way, MR authors have an easier time identifying a communication channel to approach the approver group as a whole instead of reaching out to individuals.

the slack user groups and gitlab user groups of a `RoleV1` are the relevant group communication information used for this feature.

e.g. the following message will be available within an MR

```
Reach out to approvers for reviews on
* Slack group abc
* GitLab group def
```

Ref: https://issues.redhat.com/browse/APPSRE-6847

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>