detect affected shards in desired state diff
the difference in desired state state is used to find to find the affected shards for an integration. an integration can use the `get_desired_state_shard_config` function to return a `DesiredStateShardConfig` object that defines where to find the affected shards in the desired state.

the `DesiredStateShardConfig` object can contain a callback function where the integration can decide, if sharded run should be tried based on the detected shards, e.g. this can be used to not run in sharded mode if too many shards have been detected.

this is opt int - if such a function does not exist in an integration module, or the method is not defined in the integration class, no shard detection and no sharded running is conducted.

if the difference in desired state is too big, finding the actual fine grained differences can take a very long time. therefore it makes sense to stop looking for differences after a certain timeout, becaue the gain of a sharded run would be eaten up by the diff lookup mechanism.

if we run into a timeout while looking for differences, a sharded run is not considered anymore and a regular dry-run is executed.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>