[change-owners] awareness for changes outside of the bundle
the change-owners integration writes a table of all detected bundle changes to an MR comment for give the reviewer an overview of what changes.

changes in the app-interface repo, that are not part of a bundle are not shown in this difference table. they are invisible to change-owners. but change-owners know that such changes exist (a.k.a. non-authoritative-mode) and can make the reviewer aware of such changes by leaving a note on the MR.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>