[openshift-resources] no namespaces left after cluster-filtering
when openshift-resources-base (so openshift-resource, vaults-secrets, etc) are executed with cluster or namespace filtering, there might be no namespace left after filtering. this does not necessarily indicate a usage issue but can happen during cluster deprovisioning.

this fix prevents those integrations to fail in such scenarios, but gracefully exits with a log message.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>