[integration-manager] sharding labels for integration workload (2nd try) (#3179)
the current matchLabels on a Deployment are not sufficient to catch only the pods spawned by ReplicaSets of the deployment itself, because the app label alone is the same for all potential shards of an integration.

this has no negative effect because the ReplicaSets still handle pod creation/termination correctly. the only downside is the incorrect pod filtering within the openshift UI within Deployment/Pods.

to fix this, we will add a shard label qontract_reconcile_integration_shard to both the pod template and the matchLabels of the deployment if applicable.

the first attempt of this PR failed because number based shard labels have not been cast to strings and the resulting Deployment manifest was invalid. this has been addressed in this PR and specifically covered by a test