extract runtime environment config
extract the configuration and setup of the qontract-reconcile environment to a dedicated module.
the `init_env` function sets up logging, initializes the config and sets up GQL connectivity to the `qontract-server`.

`init_env` can be called with or without parameters about loglevel and config. if no parameters are passed, their values are extracted from env variables.

if it is called with parameters, their values are stored in env variables. this way, if a child process is spawned, it can call `init_env` without parameters and a log + config + GQL environment compatible with the parent process is established.

the current way to set up the q-r env was implemented multiple times. this change unifies the process.

this is prepwork to optionally run integration shards during dry-runs with `multiprocessing.Process` instead of threading. this makes sense for certain integrations that are CPU bound and not so much I/O bound because python threading is not really allowing CPU bound tasks to benefit a lot.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>