[integration-runtime] replace integration args, kwargs run parameters with class
replace the generic `*args: Any, **kwargs: Any` parameterlist on the `run` function of class based integrations with a dedicated parameter class on an integrations `__init__` function.
wrapping parameters in a dataclass enables validations at runtime (e.g. pydantic) but also at static code analysis (e.g. mypy).

to leverage type safety all the way, the type of the parameter class can be defined as type hint when inheriting from `QontractReconcileIntegration`. the `__init__` function will then accept that type and the `param` property of the class will have that as a verifable type as well.

```python
class SimpleTestIntegrationParams(PydanticRunParams):
    int_arg: int
    opt_str_arg: Optional[str] = None

class SimpleTestIntegration(QontractReconcileIntegration[SimpleTestIntegrationParams]):
    def __init__(self, params: SimpleTestIntegrationParams):
        super().__init__(params)

    ...

    def run(self, dry_run: bool) -> None:
        print(self.params.int_arg)
```

in this example, the `print(self.params.int_arg)` statement is completely verifiable with mypy and also VSCode yields proper autocompletion.

in cli.py, the integration is hooked up like this:

```python
@integration.command()
@click.pass_context
def simple_integration(ctx, int_param: int):
    from reconcile import simple_integration

    run_class_integration(
        integration=simple_integration.SimpleIntegration(SimpleTestIntegrationParams(int_arg=int_param)),
        ctx=ctx.obj,
    )
```

for the module based integrations, the generic wrapper class `ModuleArgsKwargsRunParams` is available, that makes sure the custom `run` functions in the modules do not need to change. all the wrapping happens in `reconcile.cli` so module integrations do not need to adapt because of this change.

Wrapping all paramters in a single object makes it easier to pass them around. This comes in handy when sharded integration runs are spawned off into threads or processes.

Another option would have been to make the integration a dataclass and defining all parameters as properties. That would make it harder to distinguish between run parameters passed into the integration and potential object properties that are just there to hold some values for the integration run.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>