[change-owners] fix jsonpath sorting
for certain operations, jsonpaths need to be sorted. the currently used naive approach simply compared the jsonpath strings, which starts to fail when double digit indices are involved.

e.g. `a.b.[10] < a.b.[9]` if simple string comparison is used

this PR introduces a function that creates a sortable string representation of jsonpaths that takes care of those index numbers by padding with zeros.

https://issues.redhat.com/browse/APPSRE-7197

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>