[change-owners] fix jsonpath sorting (#3247)
for certain operations, jsonpaths need to be sorted. the currently used naive approach simply compared the jsonpath strings, which starts to fail when double digit indices are involved.

e.g. `a.b.[10] < a.b.[9]` if simple string comparison is used

this PR introduces a function that creates a sortable string representation of jsonpaths that takes care of those index numbers by padding with zeros.

this is supposed to be a temporary fix

https://issues.redhat.com/browse/APPSRE-7197