allowedSecretParameterPaths prefix but not parent of secretParameters (#3242)
trying to make a test for another example of #3227.  In this test, we have `foo` in `allowedSecretParameterPaths` but the `secretParameters` has `path = "foobar/baz"`.

I think this means validation should fail in `_validate_allowed_secret_parameter_paths` with the log message `secret parameter path 'foobar/baz' does not match any of allowedSecretParameterPaths`.

But this is based on a naive reading of the code + test cases, and I could be wrong.  Somebody who uses this function should review.

https://issues.redhat.com/browse/APPSRE-7196