Bugfix for gitlab for compliance
This fixes a bug in gitlab fork compliance that will falsly report devtools-bot missing in cases where the user was not added to the repository directly but got permission by adding the grup instead. Prominent case for this is when devtools bot itself is creating MRs directly on the repository in question, this is common for automated MRs.