Bugfix for gitlab for compliance (#3255)
* Bugfix for gitlab for compliance

This fixes a bug in gitlab fork compliance that will falsely report devtools-bot missing in cases where the user was not added to the repository directly but got permission by adding the grup instead. Prominent case for this is when devtools bot itself is creating MRs directly on the repository in question, this is common for automated MRs.

* Fix lint, remove unused import

* Add try catch for gitlaberror again

This will catch errors were i.e. the repo is private or no access was granted

* Fix format