[ocm] distinguish between OCM environment and OCM org (#3334)
`/openshift/openshift-cluster-manager-1.yml` is a schema that hybridly declares OCM environments and OCM organizations. this blures the line what parts belong to an environment and which ones to an organization.

in this PR, the schema split into an environment specific part and an org part is implemented. authentication credentials and url are present in the environment part, while the org part looses the url but keeps credentials as optional override.

schema change: https://github.com/app-sre/qontract-schemas/pull/410
see APPSRE-7281

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>