upgrade sretoolbox to 2.3.0 and adapt exception handling for tasks (#3202)
this introduces a process based parallelisation option compatible to the sretoolbox threaded module

sretoolbox 2.3.0 is not hiding sys.exit exceptions anymore from the caller of `threaded.run`. therefore we need to adapt our interpretation of exception results.

e.g. in the case of running integration shards threaded a sys.exit(0) is still fine and not an error.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>