[ocm-utils] testability and test refactor (#3393)
this PR is a first step to harmonize the httpretty mocks and fixtures for OCM testability.
`reconcile.test.ocm.conftest` defines pluggable fixtures, e.g. for `OCMBaseClient` and `OCM` that enables concise test cases and fixture setup.

not all OCM related tests have been migrated to linted/mypy-verified/lean mocked tests so far, so this is only to be considered a first step.