[jsonpath parsing] speed up jsonpath expression parsing
With more and more jsonpaths being added to existing and new change-types, parsing those has a noticeable performance impact.

There are two parsers for jsonpath expressions, a simple one and an extended one. the simple one is considerably faster and works for most of the expressions. the extended parser is required when filters are involved. For now we can detect the difference between those two types expressions by looking for a filter literal and choose the appropriate parser.

https://issues.redhat.com/browse/APPSRE-7467

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>